Create  Medical Insurance Project using Python Functions
In this project, I created a function that calculates medical insurance costs based in the values passed into the function. I also modified one of my functions so that it returns the estimated cost and the output message. lastly, I created a second function to calculate the difference between the insurance costs of any two individuals of my choosing and print a statement.